+++ docs/about/CHANGELOG.md
@@ -10,6 +10,10 @@ This file was automatically generated for [incolume.py.changelog](https://github
 ---
 
 
+## [Unreleased]	 &#8212; 	2024-05-29:
+### Added
+  - Dojo20240527 permutação python;
+
 ## [1.51.0]	 &#8212; 	2024-05-29:
 ### Added
   - Dojo 20240527 primos de 4 algarismos;
@@ -439,3 +443,4 @@ This file was automatically generated for [incolume.py.changelog](https://github
 [1.49.0]: https://github.com/incolume-jedi/coding-dojo/compare/1.48.1...1.49.0
 [1.50.0]: https://github.com/incolume-jedi/coding-dojo/compare/1.49.0...1.50.0
 [1.51.0]: https://github.com/incolume-jedi/coding-dojo/compare/1.50.0...1.51.0
+[Unreleased]: https://github.com/incolume-jedi/coding-dojo/compare/1.51.0...Unreleased